move <Slider> event monitoring into hook.
