+++ code/system.ui/react.common/src/ui/Slider/use.EventMonitor.mts
@@ -1,0 +1,68 @@
+import { useEffect, useRef } from 'react';
+import { Wrangle } from './Wrangle.mjs';
+import { DEFAULTS, useMouseState, type t } from './common';
+
+type M = React.MouseEventHandler;
+type Args = {
+  enabled?: boolean;
+  onChange?: t.SliderTrackChangeHandler;
+};
+
+/**
+ * HOOK: hook-up and manage mouse events.
+ */
+export function useEventMonitor(args: Args = {}) {
+  const { enabled = DEFAULTS.enabled, onChange } = args;
+
+  useEffect(() => {
+    return () => removeTransientEvents(); // Dispose.
+  }, []);
+
+  const removeTransientEvents = () => {
+    const detach = document.removeEventListener;
+    detach('mousemove', onMouseMove);
+    detach('mouseup', removeTransientEvents);
+    detach('selectstart', onSelectStart);
+  };
+
+  const onDown: M = (e) => {
+    if (!enabled || !ref.current) return;
+    if (onChange) {
+      // Start listening to mouse movement (drag-start).
+      const attach = document.addEventListener;
+      attach('mousemove', onMouseMove);
+      attach('mouseup', removeTransientEvents);
+      attach('selectstart', onSelectStart);
+
+      // Fire event.
+      const percent = Wrangle.elementToPercent(ref.current, e.clientX);
+      onChange({ percent });
+    }
+  };
+
+  const onMouseMove = (e: MouseEvent) => {
+    if (!enabled || !ref.current) return;
+    const percent = Wrangle.elementToPercent(ref.current, e.clientX);
+    onChange?.({ percent });
+  };
+
+  const onSelectStart = (e: Event) => {
+    // NB: Prevent content around the slider from being
+    //     selected while thumb is dragging.
+    e.preventDefault();
+  };
+
+  const ref = useRef<HTMLDivElement>(null);
+  const mouse = useMouseState({ onDown });
+
+  /**
+   * API
+   */
+  return {
+    ref,
+    el: ref.current,
+    enabled,
+    pressed: mouse.isDown,
+    handlers: mouse.handlers,
+  } as const;
+}
